scripts: set KUBEBUILDER_ASSETS dir for tests
