+++ pkg/controller/kubefedcluster/controller_test.go
@@ -17,102 +17,19 @@ limitations under the License.
 package kubefedcluster
 
 import (
-	"context"
-	"path/filepath"
+	"reflect"
 	"testing"
 	"time"
 
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/ginkgo/extensions/table"
-	. "github.com/onsi/gomega"
-
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	//schema "k8s.io/apimachinery/pkg/runtime/schema"
-	//"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	"sigs.k8s.io/controller-runtime/pkg/envtest/printer"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	fedapis "sigs.k8s.io/kubefed/pkg/apis"
 	"sigs.k8s.io/kubefed/pkg/apis/core/common"
 	fedv1b1 "sigs.k8s.io/kubefed/pkg/apis/core/v1beta1"
 	"sigs.k8s.io/kubefed/pkg/controller/util"
 )
 
-func TestController(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecsWithDefaultAndCustomReporters(t, "Controller Integration Suite", []Reporter{printer.NewlineReporter{}})
-}
-
-var testenv *envtest.Environment
-var cfg *rest.Config
-var clientset *kubernetes.Clientset
-var k8sClient client.Client
-var cc *ClusterController
-var stopControllerCh chan struct{}
-var config *util.ClusterHealthCheckConfig
-
-var _ = BeforeSuite(func(done Done) {
-	logf.SetLogger(zap.LoggerTo(GinkgoWriter, true))
-
-	testenv = &envtest.Environment{
-		CRDDirectoryPaths: []string{
-			filepath.Join("..", "testdata", "fixtures"),
-		},
-	}
-
-	var err error
-	cfg, err = testenv.Start()
-	Expect(err).NotTo(HaveOccurred())
-	Expect(cfg).ToNot(BeNil())
-
-	clientset, err = kubernetes.NewForConfig(cfg)
-	Expect(err).NotTo(HaveOccurred())
-
-	scheme := runtime.NewScheme()
-	err = fedapis.AddToScheme(scheme)
-	Expect(err).NotTo(HaveOccurred())
-
-	k8sClient, err = client.New(cfg, client.Options{Scheme: scheme})
-	Expect(err).NotTo(HaveOccurred())
-	Expect(k8sClient).ToNot(BeNil())
-
-	config = &util.ClusterHealthCheckConfig{
-		Period:           10 * time.Second,
-		FailureThreshold: 3,
-		SuccessThreshold: 1,
-		Timeout:          10 * time.Second,
-	}
-
-	controllerConfig := &util.ControllerConfig{
-		KubeFedNamespaces: util.KubeFedNamespaces{
-			KubeFedNamespace: util.DefaultKubeFedSystemNamespace,
-		},
-		KubeConfig:      cfg,
-		MinimizeLatency: true,
-	}
-
-	stopControllerCh = make(chan struct{})
-	cc, err = newClusterController(controllerConfig, config)
-	Expect(err).ToNot(HaveOccurred())
-
-	cc.Run(stopControllerCh)
-
-	close(done)
-}, 60)
-
-var _ = AfterSuite(func() {
-	close(stopControllerCh)
-	Expect(testenv.Stop()).To(Succeed())
-})
-
-var _ = Describe("TestKubefedClusterController", func() {
+func TestThresholdCheckedClusterStatus(t *testing.T) {
 
 	epoch := metav1.Now()
 	t1 := metav1.Time{Time: epoch.Add(1 * time.Second)}
@@ -121,124 +38,71 @@ var _ = Describe("TestKubefedClusterController", func() {
 	t4 := metav1.Time{Time: epoch.Add(4 * time.Second)}
 	t5 := metav1.Time{Time: epoch.Add(5 * time.Second)}
 
-	It("validate controller informer actions and cluster data functions", func() {
-		ctx := context.TODO()
-		kubefedClusterSecret := &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: util.DefaultKubeFedSystemNamespace,
-				Name:      "test-cluster-data",
-			},
-			Data: map[string][]byte{
-				util.TokenKey: []byte("xxxxx"),
-			},
-		}
+	config := &util.ClusterHealthCheckConfig{
+		Period:           10 * time.Second,
+		FailureThreshold: 3,
+		SuccessThreshold: 1,
+		Timeout:          3 * time.Second,
+	}
 
-		_, err := clientset.CoreV1().Secrets(util.DefaultKubeFedSystemNamespace).Create(kubefedClusterSecret)
-		Expect(err).ToNot(HaveOccurred())
-
-		kc := &fedv1b1.KubeFedCluster{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "cluster1",
-				Namespace: util.DefaultKubeFedSystemNamespace,
-				Labels: map[string]string{
-					"foo": "bar",
-				},
-			},
-			Spec: fedv1b1.KubeFedClusterSpec{
-				APIEndpoint: "https://my.example.com:80/path/to/endpoint",
-				SecretRef: fedv1b1.LocalSecretReference{
-					Name: kubefedClusterSecret.Name,
-				},
-			},
-			Status: *clusterStatus(corev1.ConditionTrue, metav1.Time{Time: metav1.Now().Add(1 * time.Second)}, metav1.Time{Time: metav1.Now().Add(1 * time.Second)}),
-		}
-
-		err = k8sClient.Create(ctx, kc)
-		Expect(err).ToNot(HaveOccurred())
-
-		cc.addToClusterSet(kc)
-		Expect(err).ToNot(HaveOccurred())
-		_, found := cc.clusterDataMap[kc.Name]
-		Expect(found).To(BeTrue())
-
-		cc.delFromClusterSet(kc)
-		_, found = cc.clusterDataMap[kc.Name]
-		Expect(found).NotTo(BeTrue())
-
-		fedCluster := &fedv1b1.KubeFedCluster{}
-		err = k8sClient.Get(ctx, client.ObjectKey{Namespace: util.DefaultKubeFedSystemNamespace, Name: kc.Name}, fedCluster)
-		Expect(err).ToNot(HaveOccurred())
-
-		fedCluster.Spec.APIEndpoint = "https://my.example.com:80/path/to/newendpoint"
-		err = k8sClient.Update(ctx, fedCluster)
-		Expect(err).ToNot(HaveOccurred())
-
-		Eventually(func() bool {
-			_, found = cc.clusterDataMap[kc.Name]
-			return found
-		}, 30*time.Second).Should(BeTrue())
-	})
-
-	DescribeTable(
-		"threshold checked cluster status",
-		func(clusterStatus *fedv1b1.KubeFedClusterStatus,
-			storedClusterData *ClusterData,
-			expectedClusterStatus *fedv1b1.KubeFedClusterStatus,
-			expectedResultRun int64) {
-			newClusterStatus := thresholdAdjustedClusterStatus(clusterStatus, storedClusterData, config)
-			//Expect(expectedClusterStatus).To(Equal(newClusterStatus))
-
-			//Expect(expectedResultRun).To(Equal(storedClusterData.resultRun))
-			Expect(expectedClusterStatus).To(
-				Equal(newClusterStatus),
-				"Unexpected state, expected: %v, got:%v",
-				expectedClusterStatus,
-				newClusterStatus)
-
-			Expect(expectedResultRun).To(
-				Equal(storedClusterData.resultRun),
-				"Unexpected resultRun, expected: %v, got:%v",
-				expectedResultRun,
-				storedClusterData.resultRun)
+	testCases := map[string]struct {
+		clusterStatus         *fedv1b1.KubeFedClusterStatus
+		storedClusterData     *ClusterData
+		expectedClusterStatus *fedv1b1.KubeFedClusterStatus
+		expectedResultRun     int64
+	}{
+		"ClusterReadyAtBegining": {
+			clusterStatus:         clusterStatus(corev1.ConditionTrue, t1, t1),
+			storedClusterData:     &ClusterData{clusterStatus: nil, resultRun: 0},
+			expectedClusterStatus: clusterStatus(corev1.ConditionTrue, t1, t1),
+			expectedResultRun:     1,
 		},
-		Entry("ClusterReadyAtBegining",
-			clusterStatus(corev1.ConditionTrue, t1, t1),
-			&ClusterData{clusterStatus: nil, resultRun: 0},
-			clusterStatus(corev1.ConditionTrue, t1, t1),
-			int64(1),
-		),
-		Entry("ClusterNotReadyAtBegining",
-			clusterStatus(corev1.ConditionFalse, t1, t1),
-			&ClusterData{clusterStatus: nil, resultRun: 0},
-			clusterStatus(corev1.ConditionFalse, t1, t1),
-			int64(1),
-		),
-		Entry("ClusterNotReadyButWithinFailureThreshold",
-			clusterStatus(corev1.ConditionFalse, t3, t3),
-			&ClusterData{
+		"ClusterNotReadyAtBegining": {
+			clusterStatus:         clusterStatus(corev1.ConditionFalse, t1, t1),
+			storedClusterData:     &ClusterData{clusterStatus: nil, resultRun: 0},
+			expectedClusterStatus: clusterStatus(corev1.ConditionFalse, t1, t1),
+			expectedResultRun:     1,
+		},
+		"ClusterNotReadyButWithinFailureThreshold": {
+			clusterStatus: clusterStatus(corev1.ConditionFalse, t3, t3),
+			storedClusterData: &ClusterData{
 				clusterStatus: clusterStatus(corev1.ConditionTrue, t2, t1),
 				resultRun:     2},
-			clusterStatus(corev1.ConditionTrue, t3, t1),
-			int64(3),
-		),
-		Entry("ClusterNotReadyAndCrossedFailureThreshold",
-			clusterStatus(corev1.ConditionFalse, t4, t4),
-			&ClusterData{
+			expectedClusterStatus: clusterStatus(corev1.ConditionTrue, t3, t1),
+			expectedResultRun:     3,
+		},
+		"ClusterNotReadyAndCrossedFailureThreshold": {
+			clusterStatus: clusterStatus(corev1.ConditionFalse, t4, t4),
+			storedClusterData: &ClusterData{
 				clusterStatus: clusterStatus(corev1.ConditionTrue, t3, t1),
 				resultRun:     3},
-			clusterStatus(corev1.ConditionFalse, t4, t4),
-			int64(1),
-		),
-		Entry("ClusterReturnToReadyState",
-			clusterStatus(corev1.ConditionTrue, t5, t5),
-			&ClusterData{
+			expectedClusterStatus: clusterStatus(corev1.ConditionFalse, t4, t4),
+			expectedResultRun:     1,
+		},
+		"ClusterReturnToReadyState": {
+			clusterStatus: clusterStatus(corev1.ConditionTrue, t5, t5),
+			storedClusterData: &ClusterData{
 				clusterStatus: clusterStatus(corev1.ConditionFalse, t4, t1),
 				resultRun:     1},
-			clusterStatus(corev1.ConditionTrue, t5, t5),
-			int64(1),
-		),
-	)
-})
+			expectedClusterStatus: clusterStatus(corev1.ConditionTrue, t5, t5),
+			expectedResultRun:     1,
+		},
+	}
+
+	for testName, tc := range testCases {
+		t.Run(testName, func(t *testing.T) {
+			newClusterStatus := thresholdAdjustedClusterStatus(tc.clusterStatus, tc.storedClusterData, config)
+			if !reflect.DeepEqual(tc.expectedClusterStatus, newClusterStatus) {
+				t.Fatalf("Unexpected state, expected: %v, got:%v", tc.expectedClusterStatus, newClusterStatus)
+			}
+
+			if tc.expectedResultRun != tc.storedClusterData.resultRun {
+				t.Fatalf("Unexpected resultRun, expected: %v, got:%v", tc.expectedResultRun, tc.storedClusterData.resultRun)
+			}
+		})
+	}
+
+}
 
 func clusterStatus(status corev1.ConditionStatus, lastProbeTime, lastTransitionTime metav1.Time) *fedv1b1.KubeFedClusterStatus {
 	return &fedv1b1.KubeFedClusterStatus{
